+++ db/migrate/20180405141051_add_lottery_number_uniqueness_index.rb
@@ -1,0 +1,5 @@
+class AddLotteryNumberUniquenessIndex < ActiveRecord::Migration[5.1]
+  def change
+    add_index(:lottery_assignments, [:draw_id, :number], unique: true)
+  end
+end