+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180326062759) do
+ActiveRecord::Schema.define(version: 20180405141051) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -112,6 +112,7 @@ ActiveRecord::Schema.define(version: 20180326062759) do
     t.datetime "updated_at", null: false
     t.bigint "clip_id"
     t.index ["clip_id"], name: "index_lottery_assignments_on_clip_id"
+    t.index ["draw_id", "number"], name: "index_lottery_assignments_on_draw_id_and_number", unique: true
     t.index ["draw_id"], name: "index_lottery_assignments_on_draw_id"
   end