+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Changed
+* Add restrictions to make sure lottery assignments don't exist without groups ([#780](https://gitlab.com/yale-sdmp/vesta/issues/780)).
+
 ## v1.0.6 - 2018-04-02
 ### Added
 * Add an overview for suite selection ([#723](https://gitlab.com/yale-sdmp/vesta/issues/723)).